Rename Markdown.md to major business functions
The following table lists some of the major business functions at Music Mall. The table has one type of data that might be generated and recorded for each of the business functions, A brief explanation of each type of data, and how the data could be used in a meaningful way to gain insight into business functions or to inform decision making.